removed mod.rs and removed exmample entry from cargo.toml
