+++ Cargo.toml
@@ -27,7 +27,3 @@ log = "0.4"
 failure = "0.1"
 failure_derive = "0.1"
 
-[[example]]
-name = "playground"
-path = "examples/mod.rs"
-
