+++ examples/udp.rs
@@ -10,6 +10,21 @@ use self::laminar::packet::Packet;
 
 use super::{client_address, server_address};
 
+/// The socket address of where the server is located.
+const SERVER_ADDR: &'static str = "127.0.0.1:12345";
+// The client address from where the data is sent.
+const CLIENT_ADDR: &'static str = "127.0.0.1:12346";
+
+fn client_address() -> SocketAddr
+{
+    CLIENT_ADDR.parse().unwrap()
+}
+
+fn server_address() -> SocketAddr
+{
+    SERVER_ADDR.parse().unwrap()
+}
+
 /// This is an example of how to send data to an specific address.
 pub fn send_data()
 {
